issue #46 change Currency and double to BigDecimal
edit method getCalculateSum and getTotal

+++ webcash-domain/src/main/java/com/siriusif/model/helpers/SaleBuiledr.java
@@ -1,5 +1,7 @@
 package com.siriusif.model.helpers;
 
+import java.math.BigDecimal;
+
 import com.siriusif.model.Good;
 import com.siriusif.model.Sale;
 
@@ -12,7 +14,7 @@ public class SaleBuiledr {
 	 * @param price
 	 * @return
 	 */
-	public static Sale buildSale(String name, double amount, double price) {
+	public static Sale buildSale(String name, BigDecimal amount, BigDecimal price) {
 		Sale sale = new Sale();
 		sale.setSalesGood(new Good(name, price));
 		sale.setAmount(amount);
